Drop Jade namespace
