+++ .travis.yml
@@ -15,9 +15,6 @@ matrix:
     - php: 7.2
       env:
         - COVERAGE=true
-    - php: 7.2
-      env:
-        - COVERAGE=true
     - php: 7.3
     - php: 7.4
     - php: 7.4
