+++ composer.json
@@ -32,7 +32,7 @@
   ],
   "autoload": {
     "psr-4": {
-      "Pug": "src/Pug/"
+      "Pug\\": "src/Pug/"
     }
   },
   "scripts": {
