--- src/Jade/Symfony/JadeEngine.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Jade\Symfony;
-
-use Jade\Jade;
-
-class JadeEngine extends Jade
-{
-    // Jade engine customization come here.
-}
