+++ src/Pug/Symfony/Css.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Jade\Symfony;
+namespace Pug\Symfony;
 
 use Symfony\Bundle\FrameworkBundle\Templating\Helper\AssetsHelper;
