+++ src/Pug/Symfony/Traits/PrivatePropertyAccessor.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Jade\Symfony\Traits;
+namespace Pug\Symfony\Traits;
 
 use ReflectionProperty;
