+++ tests/project/app/config.yml
@@ -1,11 +1,11 @@
 framework:
     secret: secret
     test: ~
-    router: { resource: "%kernel.root_dir%/routing.yml" }
+    router: { resource: "%kernel.project_dir%/routing.yml" }
     form: true
     csrf_protection: true
     validation: { enable_annotations: true }
-    templating: { engines: ['pug', 'php'] }
+    
     session:
         storage_id: session.storage.filesystem
